Restructured files for removal of page routing - tab bar mechanics inserted
