--- src/app/about/page.tsx
@@ -1,10 +1,0 @@
-'use client';
-
-import React from 'react';
-import About from '@/app/about/about';
-
-
-export default function AboutPage() {
-
-  return <About />
-}
