--- src/app/experience/page.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import React from 'react';
-import Experience from '@/app/experience/experience';
-
-export default function ExperiencePage() {
-
-  return <Experience />
-}
