--- src/app/works/page.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import React from 'react';
-import Works from '@/app/works/works';
-
-export default function WorksPage() {
-
-  return <Works />
-}
